Refactor: switch to folder-based routing
