+++ frontend/__tests__/routes/_app+/personal-information+/home-address+/edit.test.tsx
@@ -2,10 +2,11 @@ import { redirect } from '@remix-run/node';
 
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import { action, loader } from '~/routes/_gcweb-app.personal-information.mailing-address.edit';
+import { action, loader } from '~/routes/_app+/personal-information+/home-address+/edit';
 import { getAddressService } from '~/services/address-service.server';
 import { getSessionService } from '~/services/session-service.server';
 import { getUserService } from '~/services/user-service.server';
+import { getWSAddressService } from '~/services/wsaddress-service.server';
 
 vi.mock('~/services/address-service.server', () => ({
   getAddressService: vi.fn().mockReturnValue({
@@ -60,7 +61,13 @@ vi.mock('~/services/user-service.server', () => ({
   }),
 }));
 
-describe('_gcweb-app.personal-information.mailing-address.edit', () => {
+vi.mock('~/services/wsaddress-service.server', () => ({
+  getWSAddressService: vi.fn().mockReturnValue({
+    correctAddress: vi.fn(),
+  }),
+}));
+
+describe('_gcweb-app.personal-information.home-address.edit', () => {
   afterEach(() => {
     vi.clearAllMocks();
   });
@@ -74,7 +81,7 @@ describe('_gcweb-app.personal-information.mailing-address.edit', () => {
       vi.mocked(addressService.getAddressInfo).mockResolvedValue({ address: '111 Fake Home St', city: 'city', country: 'country' });
 
       const response = await loader({
-        request: new Request('http://localhost:3000/personal-information/mailing-address/edit'),
+        request: new Request('http://localhost:3000/personal-information/home-address/edit'),
         context: {},
         params: {},
       });
@@ -83,7 +90,6 @@ describe('_gcweb-app.personal-information.mailing-address.edit', () => {
 
       expect(data).toEqual({
         addressInfo: { address: '111 Fake Home St', city: 'city', country: 'country' },
-        homeAddressInfo: { address: '111 Fake Home St', city: 'city', country: 'country' },
         countryList: [
           {
             code: 'SUP',
@@ -111,7 +117,7 @@ describe('_gcweb-app.personal-information.mailing-address.edit', () => {
 
       try {
         await loader({
-          request: new Request('http://localhost:3000/personal-information/mailing-address/edit'),
+          request: new Request('http://localhost:3000/personal-information/home-address/edit'),
           context: {},
           params: {},
         });
@@ -122,10 +128,14 @@ describe('_gcweb-app.personal-information.mailing-address.edit', () => {
   });
 
   describe('action()', () => {
-    it('should redirect to confirm page', async () => {
+    it('should redirect to confirm page if the result of correcting the address is Valid', async () => {
       const sessionService = await getSessionService();
-
       vi.mocked(sessionService.commitSession).mockResolvedValue('some-set-cookie-header');
+
+      const wsAddressService = await getWSAddressService();
+      vi.mocked(wsAddressService.correctAddress, { partial: true }).mockResolvedValue({
+        status: 'Valid',
+      });
 
       const formData = new FormData();
       formData.append('address', '111 Fake Home St');
@@ -135,28 +145,62 @@ describe('_gcweb-app.personal-information.mailing-address.edit', () => {
       formData.append('country', 'country');
 
       const response = await action({
-        request: new Request('http://localhost:3000/personal-information/mailing-address/edit', { method: 'POST', body: formData }),
+        request: new Request('http://localhost:3000/personal-information/home-address/edit', { method: 'POST', body: formData }),
         context: {},
         params: {},
       });
 
-      expect(response).toEqual(redirect('/personal-information/mailing-address/confirm', { headers: { 'Set-Cookie': 'some-set-cookie-header' } }));
+      expect(response).toEqual(redirect('/personal-information/home-address/confirm', { headers: { 'Set-Cookie': 'some-set-cookie-header' } }));
     });
 
-    it('should throw 404 response if userInfo is not found', async () => {
-      const userService = getUserService();
+    it('should redirect to suggested address page if the result of correcting the address is Corrected', async () => {
+      const sessionService = await getSessionService();
+      vi.mocked(sessionService.commitSession).mockResolvedValue('some-set-cookie-header');
 
-      vi.mocked(userService.getUserId).mockResolvedValue('');
+      const wsAddressService = await getWSAddressService();
+      vi.mocked(wsAddressService.correctAddress, { partial: true }).mockResolvedValue({
+        status: 'Corrected',
+      });
 
-      try {
-        await loader({
-          request: new Request('http://localhost:3000/personal-information/mailing-address/edit'),
-          context: {},
-          params: {},
-        });
-      } catch (error) {
-        expect((error as Response).status).toEqual(404);
-      }
+      const formData = new FormData();
+      formData.append('address', '111 Fake Home St');
+      formData.append('city', 'city');
+      formData.append('province', 'province');
+      formData.append('postalCode', 'postalCode');
+      formData.append('country', 'country');
+
+      const response = await action({
+        request: new Request('http://localhost:3000/personal-information/home-address/edit', { method: 'POST', body: formData }),
+        context: {},
+        params: {},
+      });
+
+      expect(response).toEqual(redirect('/personal-information/home-address/suggested', { headers: { 'Set-Cookie': 'some-set-cookie-header' } }));
+    });
+
+    it('should redirect to address accuracy page if the result of correcting the address is NotCorrect', async () => {
+      const sessionService = await getSessionService();
+      vi.mocked(sessionService.commitSession).mockResolvedValue('some-set-cookie-header');
+
+      const wsAddressService = await getWSAddressService();
+      vi.mocked(wsAddressService.correctAddress, { partial: true }).mockResolvedValue({
+        status: 'NotCorrect',
+      });
+
+      const formData = new FormData();
+      formData.append('address', '111 Fake Home St');
+      formData.append('city', 'city');
+      formData.append('province', 'province');
+      formData.append('postalCode', 'postalCode');
+      formData.append('country', 'country');
+
+      const response = await action({
+        request: new Request('http://localhost:3000/personal-information/home-address/edit', { method: 'POST', body: formData }),
+        context: {},
+        params: {},
+      });
+
+      expect(response).toEqual(redirect('/personal-information/home-address/address-accuracy', { headers: { 'Set-Cookie': 'some-set-cookie-header' } }));
     });
   });
 });
