+++ frontend/__tests__/routes/_app+/personal-information+/phone-number+/confirm.test.tsx
@@ -2,7 +2,7 @@ import { redirect } from '@remix-run/node';
 
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import { action, loader } from '~/routes/_gcweb-app.personal-information.phone-number.confirm';
+import { action, loader } from '~/routes/_app+/personal-information+/phone-number+/confirm';
 import { getSessionService } from '~/services/session-service.server';
 import { getUserService } from '~/services/user-service.server';
