+++ frontend/app/utils/locale-utils.ts
@@ -4,7 +4,7 @@ import I18nextBrowserLanguageDetector from 'i18next-browser-languagedetector';
 import I18NextHttpBackend from 'i18next-http-backend';
 import { initReactI18next } from 'react-i18next';
 
-import type { SwitchLanguageData } from '~/routes/api.switch-language';
+import type { SwitchLanguageData } from '~/routes/api+/switch-language';
 import { getClientEnv } from '~/utils/env-utils';
 import { i18nNamespacesSchema } from '~/utils/route-utils';
 import type { RouteHandleData } from '~/utils/route-utils';
